もうちょい切り出した
