+++ src/view/right-sidebar/settings.tsx
@@ -1,14 +1,12 @@
 import { getStore, updateStore, useStoreValue } from "../../store/store";
 import { DEFAULT_WORKER_COUNT } from "../../store/sync-storage/settings";
-import { Slider } from "@/components/ui/slider";
 import { useState } from "react";
 import { Button } from "@/components/ui/button";
 import { readPOIListFromClipboard } from "@/store/sync-storage/poi-list";
 import { useToast } from "@/components/ui/use-toast";
 import { IconCircleCheck } from "@tabler/icons-react";
-import { prepareWorkerPool } from "@/worker-pool/worker-pool";
+import { prepareWorkerPool } from "@/worker-pool/pool-instance";
 import { ValueSlider } from "@/components/slider-wrapper";
-import { set } from "idb-keyval";
 
 const createWorkerCountValues = () => {
   const base = DEFAULT_WORKER_COUNT;
