+++ src/worker-pool/pool-instance.ts
@@ -1,5 +1,23 @@
-import { JobType } from "@/types";
-import { MandelbrotFacadeLike } from "./worker-facade";
+import { JobType, MandelbrotWorkerType } from "@/types";
+import {
+  CalcIterationWorker,
+  CalcReferencePointWorker,
+  MandelbrotFacadeLike,
+} from "./worker-facade";
+import { getStore, updateStore } from "@/store/store";
+import { clearTaskQueue } from "./task-queue";
+import { clearWorkerReference } from "./worker-reference";
+import {
+  onCalcIterationWorkerResult,
+  onCalcIterationWorkerIntermediateResult,
+  onCalcIterationWorkerProgress,
+} from "./callbacks/iteration-worker";
+import {
+  onCalcReferencePointWorkerResult,
+  onCalcReferencePointWorkerTerminated,
+  onCalcReferencePointWorkerProgress,
+} from "./callbacks/ref-orbit-worker";
+import { clearBatchContext, tickWorkerPool } from "./worker-pool";
 
 type WorkerPool = MandelbrotFacadeLike[];
 
@@ -62,3 +80,104 @@ export const calcNormalizedWorkerIndex = (
 
   return workerIdx;
 };
+
+/**
+ * WorkerPoolを再構築する
+ * countやworkerTypeが変わった場合に呼ばれる
+ */
+export function prepareWorkerPool(
+  count: number = getStore("workerCount"),
+  workerType: MandelbrotWorkerType = getStore("mode"),
+) {
+  console.debug(`prepareWorkerPool: ${count}, ${workerType}`);
+
+  updateStore("mode", workerType);
+  if (workerType === "normal") {
+    updateStore("shouldReuseRefOrbit", false);
+  }
+
+  resetWorkers();
+
+  fillCalcIterationWorkerPool(count, workerType);
+  fillCalcReferencePointWorkerPool(1 /* 仮 */, workerType);
+}
+
+/**
+ * WorkerPoolを溜まっていたJobごと全部リセットする
+ */
+export function resetWorkers() {
+  iterateAllWorker((workerFacade) => {
+    workerFacade.clearCallbacks();
+    workerFacade.terminate();
+  });
+  resetAllWorker();
+
+  clearTaskQueue();
+  clearWorkerReference();
+  clearBatchContext();
+}
+
+/**
+ * 指定した数になるまでWorkerPoolを埋める
+ */
+function fillCalcIterationWorkerPool(
+  upTo: number = getStore("workerCount"),
+  workerType: MandelbrotWorkerType = getStore("mode"),
+) {
+  let fillCount = 0;
+  const pool = getWorkerPool("calc-iteration");
+
+  for (let i = 0; pool.length < upTo && i < upTo; i++) {
+    const workerFacade = new CalcIterationWorker(workerType);
+
+    workerFacade.onResult((...args) => {
+      onCalcIterationWorkerResult(...args);
+      tickWorkerPool();
+    });
+    workerFacade.onIntermediateResult(onCalcIterationWorkerIntermediateResult);
+    workerFacade.onProgress(onCalcIterationWorkerProgress);
+
+    pool.push(workerFacade);
+
+    fillCount++;
+  }
+
+  if (fillCount > 0) {
+    console.info(
+      `Iteration Worker filled: fill count = ${fillCount}, pool size = ${pool.length}`,
+    );
+  }
+}
+
+function fillCalcReferencePointWorkerPool(
+  upTo: number = 1,
+  workerType: MandelbrotWorkerType = getStore("mode"),
+) {
+  if (workerType !== "perturbation") return;
+
+  let fillCount = 0;
+  const pool = getWorkerPool("calc-reference-point");
+
+  for (let i = 0; pool.length < upTo && i < upTo; i++) {
+    const worker = new CalcReferencePointWorker();
+
+    worker.init();
+
+    worker.onResult((...args) => {
+      onCalcReferencePointWorkerResult(...args);
+      tickWorkerPool();
+    });
+    worker.onTerminate(onCalcReferencePointWorkerTerminated);
+    worker.onProgress(onCalcReferencePointWorkerProgress);
+
+    pool.push(worker);
+
+    fillCount++;
+  }
+
+  if (fillCount > 0) {
+    console.info(
+      `RefPoint Worker filled: fill count = ${fillCount}, pool size = ${pool.length}`,
+    );
+  }
+}
